+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 4.0.0 (unreleased)
+## 4.0.0
 
 - Drop support for Python 2.7, 3.5 and 3.6
 - Add support for Python 3.10